server: refactored express app creation
